+++ character_dialogue/scripts/mods/character_dialogue/character_dialogue_data.lua
@@ -1,0 +1,8 @@
+local mod = get_mod("character_dialogue")
+
+return {
+    name = "Character Dialogue",
+    description = mod:localize("mod_description"),
+    is_togglable = true,
+    options = { widgets = {} },
+}